requirements.txt
requests==2.31.0
loguru==0.7.0
pytz==2024.2
python-dotenv==1.0.1

Flask==2.2.5
Flask-RESTful==0.3.10
SQLAlchemy==2.0.29
Flask-Cors==4.0.0
gunicorn==20.1.0
gevent==21.12.0

redis==5.0.1
psycopg2-binary==2.9.9
minio==7.1.17
celery==5.2.3

PyJWT==2.3.0